migrate render theme settings to AndroidX (fix #12223)
