+++ main/res/xml/theme_prefs.xml
@@ -1,7 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
-<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android"
-    android:title="Map Style">
+<PreferenceScreen
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:title="@string/settings_title_map_style">
     <PreferenceCategory
         android:key="@string/pref_theme_menu"
-        android:title="@string/map_theme_options" />
+        android:title="@string/map_theme_options"
+        app:iconSpaceReserved="false" />
 </PreferenceScreen>
